Make AsyncStream Sync even if the inner future is not
The AsyncStream has a few methods that that take &self so we don't
necessarily want to unconditionally impl Send for it. However, we only
access the inner future by calling Future::poll.

We can then make AsyncStream Send in a way that is more obviously safe
by wrapping the inner stream in SyncWrapper from [0]. This gets the same
effect (AsyncStream<T> is Sync) but uses only unsafe blocks that are
more obviously safe.

[0]: https://internals.rust-lang.org/t/what-shall-sync-mean-across-an-await/12020/2